% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_2_sample.R
\name{channel_2_sample}
\alias{channel_2_sample}
\title{TMT Label Design}
\usage{
channel_2_sample(masicData, label_file, ReporterConverter)
}
\arguments{
\item{masicData}{data.frame. Original MASIC Results table}

\item{label_file}{A character vector. The path to the file that links reporter ion channels to samples}

\item{ReporterConverter}{A character vector. Path to the file that has conversion between TMT Channel and Reporter Ion}
}
\value{
Saves R object to working directory containing three data.frames: isobaricTagSamples, isobaricTagReference, and
isobaricTagFractions
}
\description{
Isobaric Tag Labeling Scheme for Samples
}
\examples{
\dontrun{
channel_2_sample(masicData, label_file = "./pre_isobaricTagSamples.txt",
ReporterConverter = "./ReporterConverter.txt")
}
}
