% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_masic.R
\name{filter_masic}
\alias{filter_masic}
\title{Filter MASIC Results}
\usage{
filter_masic(masicData, s2n_min_threshold = 4,
  interference_score = 0.9)
}
\arguments{
\item{masicData}{data.frame Original MASIC Results table}

\item{s2n_min_threshold}{numeric The minimum signal to noise ratio threshold. Default is 4}

\item{interference_score}{numeric The minimum interference score threshold. Default is 0.90}
}
\value{
data.frame A filtered version of MASIC Results table
}
\description{
Filter MASIC Results by Signal to Noise Ratios and Interference Score
}
\examples{
\dontrun{
masicData_filtered <- filter_masic(masicData)
}
}
