% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_msms.R
\name{filter_msms}
\alias{filter_msms}
\title{Filter MS-GF+ Results}
\usage{
filter_msms(msgfData,
  filterString = "PepQValue < 0.005 & abs(DelM_PPM) < 10")
}
\arguments{
\item{msgfData}{data.frame Original results table downloaded from DMS.}

\item{filterString}{Character string. A string defining the threshold values to filter PSMs.}
}
\value{
data.frame A filtered version of the MS-GF+ results table.
}
\description{
Filter MS-GF+ results by PSM scores (PepQValue and |DelM_PPM|)
}
\examples{
\dontrun{
msgfData <- filter_msms(msgfData, filterString = "PepQValue < 0.005 & abs(DelM_PPM) < 10")
}
}
