% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get Job Results from DMS}
\usage{
get_data(dataPackageNumber)
}
\arguments{
\item{dataPackageNumber}{Data package ID(s) from DMS}
}
\value{
Writes two tables to tab-delimited .txt files. (MASIC and MSGF+ Results)
}
\description{
Get Job Results from DMS
}
\examples{
\dontrun{
get_data(dp1)
get_data(c(dp1, dp2, dp3))
}
}
