% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_msms_masic.R
\name{link_msms_masic}
\alias{link_msms_masic}
\title{Link MASIC and MS-GF+ Results}
\usage{
link_msms_masic(msgfData, masicData)
}
\arguments{
\item{msgfData}{data.frame}

\item{masicData}{data.frame}
}
\value{
data.frame
}
\description{
Link MASIC and MS-GF+ Results
}
\examples{
\dontrun{
quantData <- link_msms_masic(msgfData, masicData)
}
}
