% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(crossTab)
}
\arguments{
\item{crossTab}{data.frame}
}
\value{
data.frame
}
\description{
Normalize
}
\examples{
\dontrun{
crossTab <- normalize(crossTab)
}
}
