% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_coverage_filter.R
\name{protein_coverage_filter}
\alias{protein_coverage_filter}
\title{Protein Coverage Filter}
\usage{
protein_coverage_filter(msgfData, Npep = 7, fasta)
}
\arguments{
\item{msgfData}{data.frame The msgfData dataframe filtered by PSM quality scores}

\item{fasta.files}{Character A string that links the path to the protein collection used by MS-GF+}
}
\value{
data.frame
}
\description{
Filter for protein coverage
}
\examples{
\dontrun{
msgfData <- protein_coverage_filter(msgfData, Npep = 7, fasta = "path/to/protein_collection.fasta/")
}
}
