% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_inference.R
\name{protein_inference}
\alias{protein_inference}
\title{Protein Inference}
\usage{
protein_inference(msgfData)
}
\arguments{
\item{msgfData}{data.frame}
}
\value{
data.frame
}
\description{
Protein Inference
}
\examples{
\dontrun{
msgfData_filtered_final <- protein_inference(msgfData)
}
}
